Added a color scheme, adjusted color selection
Added a blue/ocean feeling color scheme. Made random number generation use the floor method of Math. As long as user's add a color to the list of colors, it will update the number generation and maintain randomness. The more schemes, the more randomness. This method does not account for duplicates in number generation but it will still cause the current color scheme to re-assign the colors.